+++ lib/io/CStreamBuffer.cpp
@@ -76,7 +76,7 @@ CStreamBuffer::pop(UInt32 n)
 	ChunkList::iterator scan = m_chunks.begin();
 	assert(scan != m_chunks.end());
 	while (scan->size() - m_headUsed <= n) {
-		n         -= scan->size() - m_headUsed;
+		n         -= (UInt32)scan->size() - m_headUsed;
 		m_headUsed = 0;
 		scan       = m_chunks.erase(scan);
 		assert(scan != m_chunks.end());
@@ -118,7 +118,7 @@ CStreamBuffer::write(const void* vdata, UInt32 n)
 	while (n > 0) {
 		// choose number of bytes for next chunk
 		assert(scan->size() <= kChunkSize);
-		UInt32 count = kChunkSize - scan->size();
+		UInt32 count = kChunkSize - (UInt32)scan->size();
 		if (count > n)
 			count = n;
