+++ lib/platform/CMSWindowsDesks.cpp
@@ -177,7 +177,7 @@ CMSWindowsDesks::resetOptions()
 void
 CMSWindowsDesks::setOptions(const COptionsList& options)
 {
-	for (UInt32 i = 0, n = options.size(); i < n; i += 2) {
+	for (UInt32 i = 0, n = (UInt32)options.size(); i < n; i += 2) {
 		if (options[i] == kOptionWin32KeepForeground) {
 			m_leaveForegroundOption = (options[i + 1] != 0);
 			LOG((CLOG_DEBUG1 "%s the foreground window", m_leaveForegroundOption ? "Don\'t grab" : "Grab"));
@@ -774,12 +774,12 @@ CMSWindowsDesks::deskThread(void* vdesk)
 			break;
 
 		case SYNERGY_MSG_FAKE_KEY:
-			keybd_event(HIBYTE(msg.lParam), LOBYTE(msg.lParam), msg.wParam, 0);
+			keybd_event(HIBYTE(msg.lParam), LOBYTE(msg.lParam), (DWORD)msg.wParam, 0);
 			break;
 
 		case SYNERGY_MSG_FAKE_BUTTON:
 			if (msg.wParam != 0) {
-				mouse_event(msg.wParam, 0, 0, msg.lParam, 0);
+				mouse_event((DWORD)msg.wParam, 0, 0, (DWORD)msg.lParam, 0);
 			}
 			break;
 
@@ -796,7 +796,7 @@ CMSWindowsDesks::deskThread(void* vdesk)
 		case SYNERGY_MSG_FAKE_WHEEL:
 			// XXX -- add support for x-axis scrolling
 			if (msg.lParam != 0) {
-				mouse_event(MOUSEEVENTF_WHEEL, 0, 0, msg.lParam, 0);
+				mouse_event(MOUSEEVENTF_WHEEL, 0, 0, (DWORD)msg.lParam, 0);
 			}
 			break;
